Update impliment_trie.cpp
This implementation is much faster and uses less memory than the previous version. Still not great in either category though (45th and 34th percentile). I still think my previous version is a better real world implementation since it's generalizable but there are obviously optimizations that could be made to both.